Move get_workspace_head function to a module in but_workspace
Maybe we can retire this in the future, but for now, remove this from the branch actions crate
