+++ crates/but-workspace/Cargo.toml
@@ -19,6 +19,7 @@ gix = { workspace = true, features = ["worktree-mutation"] }
 gitbutler-stack.workspace = true
 gitbutler-command-context.workspace = true
 gitbutler-oxidize.workspace = true
+gitbutler-cherry-pick.workspace = true
 gitbutler-project.workspace = true
 gitbutler-commit.workspace = true
 gitbutler-repo.workspace = true
